Extract the library browser and lifecycle screens into stores and views
Continues the MainWindow decomposition. The library grid's data (albums,
composers, search, per-mode sort) moves into a LibraryBrowserStore whose
collections the shell still binds to through forwarders, with a Tier-A SortState
carrying the sort tables so mode/persistence/index logic is unit-tested off the
Windows toolchain. The composer/search panes and the whole first-run + library-
management flow (welcome, unlock, libraries manager, join, approve-device,
restore-from-cloud) become their own view classes that call back into MainWindow
for the navigation it still owns (open/close/switch). Shared building blocks
(BuildCodeDisplay, CopyToClipboard, ProviderDisplayName) land in DialogPrimitives,
ClipboardHelper, and BridgeDisplay.

MainWindow keeps the window chrome, thin XAML handlers, pane visibility, and the
status/banner renderers. The moves are behavior-preserving; the store constructs
before InitializeComponent so the x:Bind Albums/Composers forwarders resolve.
