+++ bae-windows.Tests/bae-windows.Tests.csproj
@@ -50,6 +50,7 @@
     <Compile Include="../bae-windows/Stores/PlaybackPositionModel.cs" Link="PlaybackPositionModel.cs" />
     <Compile Include="../bae-windows/Stores/SyncIndicatorModel.cs" Link="SyncIndicatorModel.cs" />
     <Compile Include="../bae-windows/UpdateFlow.cs" Link="UpdateFlow.cs" />
+    <Compile Include="../bae-windows/Stores/SortState.cs" Link="SortState.cs" />
   </ItemGroup>
 
   <ItemGroup>
